requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
backports.zoneinfo;python_version<"3.9"
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.17.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.2.4
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
